Make the '-m' commandline switch more universal
The '-m' parameter of the 'make_slackware_live.sh' script  allows you
to add more packages to the Live ISO without modifying any of the script's
"SEQ_*" package sequence variables.
However, the script would prepend "pkglist:" to the argument value, which
is redundant since 'pkglist' is the default anyway.
By removing this prepend string we can now allow a syntax like this:
"-m local:foo" to add a series of packages in the subdirectory 'foo'
of the 'local' (or 'local64') directory more easily.