Remove '--no-absolute-filenames' from cpio invocations
Since cpio 2.13, this option strips '/' and '../' from symbolic and
hard links during extraction of a cpio archive like our initramfs.
This is implemented as a fix for CVE-2015-1197 but breaks the initrd.
And leaving out this option does not have the adverse effect
I always thought it would have... archive extraction is still done
into the current directory and not in the filesystem root.