Remove KDE4BASE and add DAW as a Live Variant.
Nobody used the 1 GB ISO with a stripped KDE4 anyway, so I removed it.
DAW is a 'Work In Progress': I want a Ditigal Audio Workstation as a
Live Slackware OS with Plasma5. I want to learn if Plasma5 is lean and
fast enough for realtime audio processing. Probably it also needs PAM,
which will enter the main distro soon.