iso2usb.sh: really wipe the drive first
When a Live ISO is first copied to a USB drive using 'cp' or 'dd' and
next you use iso2usb.sh to remake the USB drive into a persistent Live OS,
traces of the old ISO9660 filesystem will remain and this confuses grub
when you boot on a UEFI computer.
So we will additionally use 'wipefs' to erase traces of filesystems.

In the help text, make it clear that '-c' and '-C' accept only integer numbers
for sizes and percentages. This is caused by a bash limitation in arithmetics.