XFCE: use the new XFCE release in testing/vtown
The increased size of applications forced a decision to remove
all development tools (glibc, gcc, make, bison, lex, llvm and the like)
from the small XFCE image. In turn I was able to re-introduce a couple of
programs that I had to prune during the years. To be honest, I have never
used any of these development tools in the XFCE Live OS anyway.
All removed development tools are now available in a separate package list
called 'z01_swdev.lst' but this module is not built by default.