make_slackware_live.sh: expand the '-t' parameter functionality
Add a new value 'bloat'.  Possible values are now:
  -t <none|doc|mandoc|bloat>
With the documentation:
  Trim the ISO (remove man and/or doc and/or bloat).
The 'value 'doc' will cause most of the documentation to be removed
(READMEs and LICENSE files will be retained). Using 'mandoc' will not just
remove documentation but additionally all man pages from the ISO.
The value 'bloat' removes a lot more, like static libraries, big shared
libraries, un-needed binaries etc. This is what is used by default when the
'XFCE' live variant is built to ensure that the ISO size stays below 700MB.

Note that from now on, you can build a XFCE ISO that has *nothing* pruned
at all! This will increase the size of the ISO, but if you were not going
to burn the ISO to a CDROM medium then that ISO size is not relevant anyway.

Example commandline:
 # make_slackware_live.sh -d XFCE -t none -c zstd
This will generate a XFCE ISO without any 'bloat' removed, and using 'zstd'
compression instead of 'xz' which will increase the size with an additional
10% but will cause the Live OS to boot significantly faster due to faster
decompression speeds of zstd.