Implement a small Console OS - a rescue environment in RAM
This uses two squashfs modules that are currently only found in the
LEAN and XFCE images: 'min' and 'noxbase'.
These two provide a functional console-only Slackware with a lot of useful
programs. It will connect to a DHCP server automatically and it also
contains the 'setup2hd' script to be able to install Slackware from a network
mirror.
And since the Console OS gets loaded into RAM, you can remove your USB stick
after booting and use that stick for other purposes.

Use-case:
- You have one computer with a network connection and one USB stick, and want
  to create a persistent Slackware Live on USB.
- Download an ISO supporting 'Console OS' to the computer's hard drive, and
  transfer the ISO to the USB stick using the computer's ISO imaging tools,
  making the stick bootable but not persistent.
- Boot from the USB stick, select the "Console OS in RAM" option.
- After you logged into the Console Slackware, mount the computer's hard drive.
- Use the 'iso2usb.sh' script that comes with liveslak to extract the ISO
  content to the USB stick, making it persistent. See the README.txt for
  instructions.

-----------------------------------------------------------------------------

Implemented as an extension of the liveslak 'toram' boot parameter.
Adding 'toram=core' to the boot commandline will load circa 500 MB of squashfs
modules into RAM and boot into a sparse but functional console environment.

For supported Live variants (currently LEAN and XFCE) the script
'make_slackware_live.sh' will automatically add a menu item "Console OS in RAM"
to the Syslinux and Grub bootloaders, using this 'toram=core' parameter.

TODO: add this as an option to all liveslak variants. Not so trivial to do.