liveinit.tpl: add support for persistence on local non-live harddisks
The syntax for the 'persistence' boot parameter has been extended.

The persistence directory/container no longer needs to be in the root
of the filesystem; you can specify an absolute path.

Additionally, you can specify the partition containing the filesystem
on which the persistence is located, or simply specify 'scandev',
to request that liveslak tries to find the partition for you:
  persistence=/dev/sdX:/path/to/mypersistence
  persistence=scandev:/path/to/mypersistence
In addition, a UUID or LABEL value of the filesystem will be accepted:
  persistence=cd68b6f5-5b5a-4d27-9649-7827489f94a5:/path/to/mypersistence

This functionality will be useful if you need the persistence directory
or container to be located somewhere else than on the Live USB stick,
for instance on a local hard disk partition.
Useful for network (PXE) boot because the overlayfs does not support a
writable upper layer of your live filesystem over NFS. If you still want
to offer persistence to users of PXE-booted Slackware Live, you can now
use the local harddisk partitions or a local USB stick for persistence.