liveinit: start ash bash-compatible and save init debug output to file
When the value of 'debug' parameter is equal to or greater than '2',
the init script enables 'set -x' which results in verbose output of
init script execution. When the 'ash' shell is started as '/bin/sh' instead
of '/bin/ash', it enables bash-compatible mode. Then the script can use
'BASH_XTRACEFD' to log the verbose script output to a file '/debug_init.log'.

This is extremely useful for init debugging. Logging to file is enabled when
the value of 'debug' is '4' or higher.
When you boot with 'debug=4, you will end up in the rescue shell of the initrd.
You can then inspect '/debug_init.log' to diagnose issues with the boot-up,
and you can copy that logfile for instance to /mnt/overlay/root/ if you want
to preserve it. After typing 'exit', the live init will hand-over to
the Slackware OS initialization and after logging in you can copy the logfile
to a network location if you want.