live init: optionally specify RAM % to use in the 'toram' parameter
toram=
  Generic syntax: toram=type[,memperc]
  - type: string value; os,core,all,none
  - memperc: integer value, percentage RAM to reserve for liveslak

You can use a 'memperc' value for the 'toram' parameter to change
the percentage RAM used by liveslak, which is 50% of all available RAM
for normal operation.
For instance when you have 128 GB of RAM, you can specify a much lower
percentage to be reserved for liveslak. RAM will specifically be used
if you run liveslak without persistence, which means that any changes are
accumulating in RAM instead of on-disk.
Example:
Give liveslak ~6 GB (5%) to run the OS on a computer with 128 GB of RAM:
  'toram=none,5'