Full support for Ventoy; encrypted OS persistence/homedirectory for ISO boot
New features:

Ventoy is now fully supported.
Liveslak implements the "Ventoy-compatible" guideline. There is nothing you
have to do, this works out of the box.

When you are booting from an ISO file (whether via Ventoy, or through your own
hand-crafted Grub menu entry, or Windows BCD), Operating System persistence
and an encrypted homedirectory are now supported, as well as the ability
to load additional live modules (as 'addons' or 'optional') that are not
part of the ISO.

All of this is possible without the need for any modification to the ISO image.
The configuration is stored in a file next to the ISO.

A new script, "isocomp.sh" aka the ISO Companion script, has been added.
Like with all of my scripts, it accepts a "--help" parameter which will
show you how to use it.
This script allows you to:
- create encrypted containers for OS persistence and a persistent homedirectory
  (actually not just for /home but you can create as many containers as
  fit on the disk and mount them wherever you want)
- size extension of existing encrypted containers if they threaten to run
  out of space
- creating a secondary liveslak root on the disk partition where you can
  add more (optional/add-on) live modules that you need in the Live OS
  but are not contained in the ISO

The configuration of these new features is stored in a file with the same
name and full path as the ISO file but with a ".cfg" file extension instead
of ".iso". The "isocomp.sh" script manages this configuration file for you,
but you can safely edit and modify it manually if you want to. The "isocomp.sh"
script will leave your customizations alone.

Here is an example of such a configuration file; it is copied from
my own Ventoy disk, for a Slackware LEAN Live ISO image:

LIVESLAKROOT=/liveslak
LUKSVOL=/liveslak/myhome.icc:/home
ISOPERSISTENCE=/liveslak/persistence.icc
TZ=Europe/Amsterdam
LIVE_HOSTNAME=zelazny

I added the variables "TZ" and "LIVE_HOSTNAME" manually by opening the
configuration file in an editor.

The following variables are supported in the isocomp configuration file,
but not managed by "isocomp.sh"; they all correspond to liveslak
boot parameters by the way:
  BLACKLIST, KEYMAP, LIVE_HOSTNAME,  LOAD, LOCALE,
  NOLOAD, RUNLEVEL, TWEAKS, TZ and XKB.

The value of the "LUKSVOL" variable can hold multiple
"containerfile:mountpoint" definitions, separated by commas.

The variable "LIVESLAKROOT" defines the root of a secondary liveslak
directory tree on your disk.