iso2usb.sh: change EFI partition label from 'EFI' to 'ESP'
Liveslak used to assign the label 'EFI' to the FAT filesystem on the
EFI System Partition (ESP). Some UEFI firmwares however cannot cope
with a volume- or labelname that is identical to the EFI directory name.
When that happens, the buggy firmware decides that an EFI directory
does not exist. The computer will not boot.
To make liveslak work on those buggy firmwares, it needs to use any name
*except* 'EFI'. Let's pick 'ESP' instead, which is still informative
about the 'EFI System Partition'.