Make kdesu work again with sudo backend (using live password)
For kdesu to work with the sudo backend in KDE5 with newer
versions of sudo (since 2022), we need to disable use_pty for kdesu.
KDE5 fixed this in git on 04-aug-2023, see
https://bugs.kde.org/show_bug.cgi?id=452532 but it does not hurt
to implement the workaround in liveslak, as it helps to support
older KDE releases.

This is the KDE commit message, see https://invent.kde.org/frameworks/kdesu/-/commit/732dd812d67c7fa62bd187c1171950ca85259b0b :

Recent versions of sudo have use_pty enabled by default, which means that sudo
creates a PTY for starting the user process inside after successful
authentication. This PTY inherits the configuration of sudo's TTY, but later
changes are not transferred. Make sure that echo is already disabled when sudo
is started, as disabling it later has no effect on the nested PTY.