Fix broken support for Ventoy on UEFI computers
Changes in Slackware-current in March and May 2023 broke support for
detection of Ventoy disk on UEFI computers.
Liveslak's initrd needed to have efivarfs module added, so that
/sys/firmware/efi/efivars can be mounted by the live init script.
With kernel 6.x the content of what's found under efivars
has changed - instead of being represented as a directory with the
efi datastructures as files, liveslak now needs to deal with
a single file instead and find the offset in there.