Fix UEFI boot when ISO was dd-ed to USB thumbdrive
Since early 2019, when you dd-ed or cp-ed a live ISO directly to a USB stick,
and you booted from this stick on a UEFI computer, you would end up in the
GRUB commandline, because $prefix would not be defined and GRUB would be
unable to find its files.
My 2019 modification fixed PXE boot but broke this specific case of UEFI boot.
And now it has finally been fixed after I had a brain fart.